+++ tests/fixtures/crates/build_dir/src/lib.rs
@@ -1,0 +1,11 @@
+fn func(x: i32) -> bool {
+    if x < 0 { true } else { false }
+}
+
+#[test]
+fn test() {
+    #[cfg(a)]
+    assert!(!func(1));
+    #[cfg(not(a))]
+    assert!(func(-1));
+}